+++ udata/tests/site/test_site_metrics.py
@@ -2,10 +2,9 @@ import pytest
 
 from udata.core.dataset.factories import DatasetFactory, VisibleDatasetFactory, OrganizationFactory
 from udata.core.reuse.factories import VisibleReuseFactory
+from udata.harvest.tests.factories import HarvestSourceFactory
 from udata.core.site.factories import SiteFactory
-from udata.models import Site, Badge, PUBLIC_SERVICE
-from udata.core.site.models import current_site
-from udata.tests.helpers import assert_emit
+from udata.models import Badge, PUBLIC_SERVICE
 
 
 @pytest.mark.usefixtures('clean_db')
@@ -68,3 +67,13 @@ class SiteMetricTest:
         site.count_org_for_badge(PUBLIC_SERVICE)
 
         assert site.get_metrics()[PUBLIC_SERVICE] == len(public_services)
+
+    def test_harvesters_metric(self, app):
+        site = SiteFactory.create(
+            id=app.config['SITE_ID']
+        )
+        sources = [HarvestSourceFactory() for i in range(10)]
+
+        site.count_harvesters()
+
+        assert site.get_metrics()['harvesters'] == len(sources)